Change Directory Tree TypedNumeric: to src/numeric/ TypedVector: to src/vector/
