+++ mod.ts
@@ -1,23 +1,23 @@
-export { Int8 } from "./src/int8.ts";
-export { Int16 } from "./src/int16.ts";
-export { Int32 } from "./src/int32.ts";
-export { Int64 } from "./src/int64.ts";
-export { Int128 } from "./src/int128.ts";
-export { Int256 } from "./src/int256.ts";
+export { Int8 } from "./src/numeric/int8.ts";
+export { Int16 } from "./src/numeric/int16.ts";
+export { Int32 } from "./src/numeric/int32.ts";
+export { Int64 } from "./src/numeric/int64.ts";
+export { Int128 } from "./src/numeric/int128.ts";
+export { Int256 } from "./src/numeric/int256.ts";
 
-export { Uint8 } from "./src/uint8.ts";
-export { Uint16 } from "./src/uint16.ts";
-export { Uint32 } from "./src/uint32.ts";
-export { Uint64 } from "./src/uint64.ts";
-export { Uint128 } from "./src/uint128.ts";
-export { Uint256 } from "./src/uint256.ts";
+export { Uint8 } from "./src/numeric/uint8.ts";
+export { Uint16 } from "./src/numeric/uint16.ts";
+export { Uint32 } from "./src/numeric/uint32.ts";
+export { Uint64 } from "./src/numeric/uint64.ts";
+export { Uint128 } from "./src/numeric/uint128.ts";
+export { Uint256 } from "./src/numeric/uint256.ts";
 
-export { Int8Vector } from "./src/int8vector.ts";
-export { Int16Vector } from "./src/int16vector.ts";
-export { Int32Vector } from "./src/int32vector.ts";
-export { Int64Vector } from "./src/int64vector.ts";
+export { Int8Vector } from "./src/vector/int8vector.ts";
+export { Int16Vector } from "./src/vector/int16vector.ts";
+export { Int32Vector } from "./src/vector/int32vector.ts";
+export { Int64Vector } from "./src/vector/int64vector.ts";
 
-export { Uint8Vector } from "./src/uint8vector.ts";
-export { Uint16Vector } from "./src/uint16vector.ts";
-export { Uint32Vector } from "./src/uint32vector.ts";
-export { Uint64Vector } from "./src/uint64vector.ts";
+export { Uint8Vector } from "./src/vector/uint8vector.ts";
+export { Uint16Vector } from "./src/vector/uint16vector.ts";
+export { Uint32Vector } from "./src/vector/uint32vector.ts";
+export { Uint64Vector } from "./src/vector/uint64vector.ts";
