+++ src/numeric/int16.ts
