+++ src/numeric/uint16.ts
