+++ src/numeric/uint256_test.ts
@@ -1,4 +1,4 @@
-import { assertEquals, assertThrows } from "../depends.ts";
+import { assertEquals, assertThrows } from "../../depends.ts";
 
 import { Uint256 } from "./uint256.ts";
