+++ src/numeric/uint32.ts
