+++ src/numeric/uint8.ts
