+++ src/numeric/uint8_test.ts
@@ -1,4 +1,4 @@
-import { assertEquals, assertThrows } from "../depends.ts";
+import { assertEquals, assertThrows } from "../../depends.ts";
 
 import { Uint8 } from "./uint8.ts";
